python-threat-intel/requirements.txt
# HTTP requests for API calls
requests>=2.31.0

# Data manipulation and analysis (updated for Python 3.13)
pandas>=2.2.0
numpy>=1.26.0

# Machine Learning for anomaly detection
scikit-learn>=1.4.0

# Command-line interface
click>=8.1.0

# Pretty terminal output
colorama>=0.4.6
rich>=13.7.0

# Environment variables
python-dotenv>=1.0.0

# Date/time handling
python-dateutil>=2.8.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

